# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BidSure Platform API
  version: 1.0.0
  description: |
    BidSure Platform API Contract — Phase 01 Architecture Baseline.
    
    ### System Conventions:
    - **API Prefix**: All endpoints are mounted under `/api/v1`.
    - **Request IDs**: Authoritative IDs generated by backend with format `req_<uuid>` (e.g., `req_550e8400-e29b-41d4-a716-446655440000`).
    - **Correlation IDs**: Optional client/system tracing identifier with format `corr_<uuid>`.
    - **Database IDs**: UUID v4 strings (frozen for Phase 03+).
    - **Timestamps**: Strictly ISO 8601 UTC format (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - **Envelope Standard**: All responses return standardized envelopes (`ApiResponseEnvelope` for 2xx, `ApiErrorEnvelope` for 4xx/5xx).
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
  - url: /api/v1
    description: Relative Path (Vite Dev Proxy)
paths:
...
  /tenders:
    get:
      summary: List Solicitations
      description: Returns paginated tenders with deterministic sorting and filtering.
      operationId: listTenders
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, CLOSED, ARCHIVED]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
      responses:
        '200':
          description: Paginated tenders list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TendersListResponse'
    post:
      summary: Create Tender
      description: Creates a new tender with initial Version 1 snapshot atomically.
      operationId: createTender
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenderRequest'
      responses:
        '201':
          description: Tender created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
        '400':
          description: Validation error.
        '409':
          description: Duplicate reference number conflict.
  /tenders/{id}:
    get:
      summary: Get Tender Detail
      description: Returns canonical tender with current version summary.
      operationId: getTenderById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tender detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
        '404':
          description: Tender not found.
    patch:
      summary: Update Tender Metadata
      description: Updates mutable administrative metadata.
      operationId: updateTenderMetadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenderMetadataRequest'
      responses:
        '200':
          description: Tender metadata updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 5
...
    TenderDTO:
      type: object
      required:
        - id
        - organizationId
        - title
        - referenceNumber
        - status
        - currentVersionId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        title:
          type: string
        referenceNumber:
          type: string
        category:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CLOSED, ARCHIVED]
        currentVersionId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentVersion:
          $ref: '#/components/schemas/TenderVersionDTO'
          nullable: true
        versionsCount:
          type: integer
    TenderVersionDTO:
      type: object
      required:
        - id
        - tenderId
        - versionNumber
        - title
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        tenderId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: object
          nullable: true
        changeSummary:
          type: string
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, SUPERSEDED]
        createdAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/RequirementDTO'
    RequirementDTO:
      type: object
      required:
        - id
        - identifier
        - category
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        identifier:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        currentVersion:
          type: object
          properties:
            versionNumber:
              type: integer
            description:
              type: string
            mandatory:
              type: boolean
    CreateTenderRequest:
      type: object
      required:
        - title
        - referenceNumber
      properties:
        title:
          type: string
        referenceNumber:
          type: string
        description:
          type: string
        category:
          type: string
        department:
          type: string
        submissionDeadline:
          type: string
          format: date-time
        budget:
          type: number
        content:
          type: object
    UpdateTenderMetadataRequest:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        department:
          type: string
        expectedVersion:
          type: integer
...
    TenderResponse:
      type: object
      required:
        - success
        - data
        - requestId
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/TenderDTO'
        requestId:
          type: string
    TendersListResponse:
      type: object
      required:
        - success
        - data
        - meta
        - requestId
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/TenderDTO'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
...